+++ src/main/java/com/quostomize/quostomize_be/domain/auth/entity/Member.java
@@ -92,9 +92,33 @@ public class Member extends BaseTimeEntity {
         this.secondaryAuthCode = secondaryAuthCode;
     }
 
-    public void updatePhoneNumber(String memberPhoneNumber) {
-        this.memberPhoneNumber = memberPhoneNumber;
+    public void updatePhoneNumber(String newPhoneNumber) {
+        if (!this.memberPhoneNumber.equals(newPhoneNumber)) {
+            this.memberPhoneNumber = newPhoneNumber;
+        }
     }
 
+    public void updateZipCode(String newZipCode) {
+        if (!this.memberAddress.equals(newZipCode)) {
+            this.memberAddress = newZipCode;
+        }
+    }
 
+    public void updateAddress(String newAddress) {
+        if (!this.memberAddress.equals(newAddress)) {
+            this.memberAddress = newAddress;
+        }
+    }
+
+    public void updateDetailAddress(String newDetailAddress) {
+        if (!this.memberDetailAddress.equals(newDetailAddress)) {
+            this.memberDetailAddress = newDetailAddress;
+        }
+    }
+
+    public void updateEmail(String newEmail) {
+        if (!this.memberEmail.equals(newEmail)) {
+            this.memberEmail = newEmail;
+        }
+    }
 }
