+++ src/main/java/com/quostomize/quostomize_be/domain/auth/repository/MemberRepository.java
@@ -12,6 +12,8 @@ import java.util.List;
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByMemberId(long memberId);
+
     Optional<Member> findByMemberLoginId(String loginId);
 
     Optional<Member> findByMemberEmail(String email);
